sql_query split
