+++ mindsdb/api/executor/__init__.py
@@ -1,0 +1,2 @@
+
+# from .sql_query.sql_query import SQLQuery, Column
