+++ mindsdb/api/executor/datahub/datanodes/information_schema_datanode.py
@@ -1,10 +1,9 @@
 from functools import partial
 
 import pandas as pd
-from mindsdb_sql.parser.ast import BinaryOperation, Constant, Identifier, Select
+from mindsdb_sql.parser.ast import BinaryOperation, Constant, Identifier, Select, Join, Union, Insert, Delete
 from mindsdb_sql.parser.ast.base import ASTNode
 
-from mindsdb.api.executor.sql_query import get_all_tables
 from mindsdb.api.executor.datahub.classes.tables_row import (
     TABLES_ROW_TYPE,
     TablesRow,
@@ -25,6 +24,33 @@ from mindsdb.interfaces.skills.skills_controller import SkillsController
 from mindsdb.utilities import log
 
 logger = log.getLogger(__name__)
+
+
+def get_all_tables(stmt):
+    if isinstance(stmt, Union):
+        left = get_all_tables(stmt.left)
+        right = get_all_tables(stmt.right)
+        return left + right
+
+    if isinstance(stmt, Select):
+        from_stmt = stmt.from_table
+    elif isinstance(stmt, (Identifier, Join)):
+        from_stmt = stmt
+    elif isinstance(stmt, Insert):
+        from_stmt = stmt.table
+    elif isinstance(stmt, Delete):
+        from_stmt = stmt.table
+    else:
+        # raise SqlApiException(f'Unknown type of identifier: {stmt}')
+        return []
+
+    result = []
+    if isinstance(from_stmt, Identifier):
+        result.append(from_stmt.parts[-1])
+    elif isinstance(from_stmt, Join):
+        result.extend(get_all_tables(from_stmt.left))
+        result.extend(get_all_tables(from_stmt.right))
+    return result
 
 
 class InformationSchemaDataNode(DataNode):
