+++ mindsdb/api/mysql/mysql_proxy/executor/mysql_executor.py
@@ -2,7 +2,7 @@ from mindsdb_sql import parse_sql
 from mindsdb_sql.planner import utils as planner_utils
 
 import mindsdb.utilities.profiler as profiler
-from mindsdb.api.executor.sql_query import Column, SQLQuery
+from mindsdb.api.executor import Column, SQLQuery
 from mindsdb.api.executor.command_executor import ExecuteCommands
 from mindsdb.api.mysql.mysql_proxy.utilities import SqlApiException
 from mindsdb.utilities import log
